merges macros.Exprs and macros.TypeTags into Context
The former is a one-method trait, the latter is a two-method trait.
In a scaladoc these guys don't look like the pull their weight.
