+++ src/reflect/scala/reflect/macros/Aliases.scala
@@ -16,11 +16,14 @@ trait Aliases {
 
   type Expr[+T] = universe.Expr[T]
   val Expr = universe.Expr
+  def Expr[T: WeakTypeTag](tree: Tree): Expr[T]
 
   type WeakTypeTag[T] = universe.WeakTypeTag[T]
   type TypeTag[T] = universe.TypeTag[T]
   val WeakTypeTag = universe.WeakTypeTag
   val TypeTag = universe.TypeTag
+  def WeakTypeTag[T](tpe: Type): WeakTypeTag[T]
+  def TypeTag[T](tpe: Type): TypeTag[T]
   def weakTypeTag[T](implicit attag: WeakTypeTag[T]) = attag
   def typeTag[T](implicit ttag: TypeTag[T]) = ttag
   def weakTypeOf[T](implicit attag: WeakTypeTag[T]): Type = attag.tpe
