--- src/reflect/scala/reflect/macros/Exprs.scala
@@ -1,8 +1,0 @@
-package scala.reflect
-package macros
-
-trait Exprs {
-  self: Context =>
-
-  def Expr[T: WeakTypeTag](tree: Tree): Expr[T]
-}
