--- src/reflect/scala/reflect/macros/TypeTags.scala
@@ -1,9 +1,0 @@
-package scala.reflect
-package macros
-
-trait TypeTags {
-  self: Context =>
-
-  def WeakTypeTag[T](tpe: Type): WeakTypeTag[T]
-  def TypeTag[T](tpe: Type): TypeTag[T]
-}
